Add version selection to .NET and Java documentation.
Version selection has been added to footer for the 2b release but in order for it to be functional and truly allow switching back and forth between releases I need to backport
support for the version selection in the previous releases.

The P1z and 2a releases were built from the source code circa the right time period but with the addition of the version selection.